Update : FriendlyUnitData 부모 스크립트와 분리
